made a hook for the notes subfolder maping
