+++ src/Components/MapRepoFolder.jsx
@@ -1,24 +1,7 @@
 import React, { useEffect, useState } from 'react';
-
+import useNoteSubFolder from '../Hooks/useNoteSubfolder';
 const MapRepoFolder = () => {
-  const [repoStructure, setRepoStructure] = useState([]);
-
-  useEffect(() => {
-    const fetchRepoData = async () => {
-      try {
-        const results = await window.getRepoPath.getRepoPath();
-        if (results.success) {
-          setRepoStructure(results.data);
-          console.log(results.data);
-        } else {
-          console.log('Error fetching repo:', results.error);
-        }
-      } catch (e) {
-        console.log('Bruh: ', e);
-      }
-    };
-    fetchRepoData();
-  }, []);
+  const repoStructure = useNoteSubFolder();
   return (
     <div>
       {repoStructure?.map((item) => (
